fix rsa lib not being added to the collection as a variable
